pyproject.toml
[project]
name = "evolsynth-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-community>=0.3.16",
    "langchain-openai>=0.3.3",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.69",
    "numpy>=2.2.2",
    "ragas==0.2.10",
    "unstructured>=0.14.8",
    "nltk==3.8.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "fastapi>=0.115.0",
    "fastjsonschema==2.21.1",
    "pydantic-ai>=0.1.1",
    "uvicorn>=0.32.0",
]
